Move repos into /server
